online_judge_backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
psutil
aio_pika
aioboto3
prometheus-client
